Refactor admin and superadmin management functions; consolidate request handling into a single file for better maintainability. Remove deprecated functions and triggers, and implement token validation middleware for enhanced security. Update user role management logic and ensure proper error handling across all functions. Add token retrieval functionality in the frontend for admin actions.
