+++ functions/src/index.ts
@@ -1,9 +1,9 @@
 /**
  * Cloud Functions for GDG Website
- *
- * This file re-exports all Cloud Functions organized in their respective folders:
- * - requests/: HTTP Cloud Functions (onRequest)
- * - triggers/: Event-triggered Cloud Functions (onDocumentWritten, beforeUserCreated)
+ * 
+ * This file re-exports all Cloud Functions organized in their respective files:
+ * - requests.ts: HTTP Cloud Functions (onRequest)
+ * - triggers.ts: Event-triggered Cloud Functions (onDocumentWritten, beforeUserCreated)
  */
 import * as admin from "firebase-admin";
 
@@ -13,15 +13,19 @@ if (!admin.apps.length) {
 }
 
 // HTTP Request Functions (Admin Operations)
-export { checkAdminClaims } from "./requests/admin/checkAdminClaims";
-export { getUsers } from "./requests/admin/getUsers";
-export { grantAdmin } from "./requests/admin/grantAdmin";
-export { grantSuperAdmin } from "./requests/admin/grantSuperAdmin";
-export { removeAdmin } from "./requests/admin/removeAdmin";
-export { addUserToTeam } from "./requests/admin/addUserToTeam";
+export {
+  checkAdminClaims,
+  getUsers,
+  grantAdmin,
+  grantSuperAdmin,
+  removeAdmin,
+  addUserToTeam
+} from "./requests";
 
 // Trigger Functions (Event-driven)
-export { beforecreated} from "./triggers/beforecreated";
-export { createApplication } from "./triggers/createApplication";
-export { createRegistration } from "./triggers/createRegistration";
-export { createCollaboration } from "./triggers/createCollaboration";
+export {
+  beforecreated,
+  createApplication,
+  createRegistration,
+  createCollaboration
+} from "./triggers";
