--- functions/src/triggers/createApplication/index.ts
@@ -1,24 +1,0 @@
-import * as admin from "firebase-admin";
-import { FieldValue } from "firebase-admin/firestore";
-import * as logger from "firebase-functions/logger";
-import { onDocumentWritten } from "firebase-functions/v2/firestore";
-
-export const createApplication = onDocumentWritten(
-  "positions/{positionId}/applications/{applicationId}",
-  async (event: any) => {
-    const positionId = event.params.positionId;
-    const userId = event.params.applicationId;
-
-    try {
-      await admin
-        .firestore()
-        .collection("users")
-        .doc(userId)
-        .update({
-          "associations.applications": FieldValue.arrayUnion(positionId),
-        });
-    } catch (error) {
-      logger.error(`Error creating application for user ${userId}:`, error);
-    }
-  }
-);
