--- functions/src/triggers/createCollaboration/index.ts
@@ -1,24 +1,0 @@
-import * as admin from "firebase-admin";
-import { FieldValue } from "firebase-admin/firestore";
-import * as logger from "firebase-functions/logger";
-import { onDocumentWritten } from "firebase-functions/v2/firestore";
-
-export const createCollaboration = onDocumentWritten(
-  "projects/{projectId}/collaborations/{collaborationId}",
-  async (event: any) => {
-    const projectId = event.params.projectId;
-    const userId = event.params.collaborationId;
-
-    try {
-      await admin
-        .firestore()
-        .collection("users")
-        .doc(userId)
-        .update({
-          "associations.collaborations": FieldValue.arrayUnion(projectId),
-        });
-    } catch (error) {
-      logger.error(`Error creating collaboration for user ${userId}:`, error);
-    }
-  }
-);
